refactor: unify tour pages structure and extract shared components
