--- src/components/pages/group-tours/Banner/styles.module.css
@@ -1,261 +1,0 @@
-.slideCounter {
-  font-family: var(--font-genseki-gothic);
-  font-size: 1.25rem;
-  line-height: 1.75rem;
-  color: var(--color-figma-primary-0);
-  display: flex;
-  align-items: center;
-  gap: 2px;
-}
-
-@media (max-width: 1279px) {
-  .slideCounter {
-    color: var(--color-figma-primary-400);
-  }
-}
-
-.controls {
-  position: absolute;
-  z-index: 10;
-  bottom: 1rem;
-  right: 50%;
-  transform: translateX(50%);
-  background-color: transparent;
-}
-
-@media (max-width: 1279px) {
-  .controls {
-    bottom: 0;
-    right: 0;
-    transform: none;
-    background-color: var(--color-figma-neutral-50);
-    border-top-left-radius: 1rem;
-  }
-
-  .controlsContainer::before {
-    content: '';
-    width: 24px;
-    height: 24px;
-    display: block;
-    position: absolute;
-    background-color: var(--color-figma-neutral-50);
-    left: -24px;
-    bottom: 0;
-    mask:
-      linear-gradient(
-        to left,
-        var(--color-figma-neutral-50) 50%,
-        transparent 50%
-      ),
-      linear-gradient(
-        to bottom,
-        transparent 50%,
-        var(--color-figma-neutral-50) 50%
-      ),
-      radial-gradient(
-        circle 12px at center,
-        transparent 12px,
-        var(--color-figma-neutral-50) 12px
-      );
-    mask-composite: intersect;
-  }
-
-  .controlsContainer::after {
-    content: '';
-    width: 24px;
-    height: 24px;
-    display: block;
-    position: absolute;
-    background-color: var(--color-figma-neutral-50);
-    right: 0;
-    top: -24px;
-    mask:
-      linear-gradient(
-        to right,
-        transparent 50%,
-        var(--color-figma-neutral-50) 50%
-      ),
-      linear-gradient(
-        to top,
-        var(--color-figma-neutral-50) 50%,
-        transparent 50%
-      ),
-      radial-gradient(
-        circle 12px at center,
-        transparent 12px,
-        var(--color-figma-neutral-50) 12px
-      );
-    mask-composite: intersect;
-  }
-}
-
-.titleContainer {
-  position: relative;
-}
-
-.titleContainer::before {
-  content: '';
-  width: 32px;
-  height: 32px;
-  display: block;
-  position: absolute;
-  background-color: var(--color-figma-neutral-50);
-  left: 0;
-  bottom: -32px;
-  mask:
-    linear-gradient(
-      to left,
-      transparent 50%,
-      var(--color-figma-neutral-50) 50%
-    ),
-    linear-gradient(
-      to bottom,
-      var(--color-figma-neutral-50) 50%,
-      transparent 50%
-    ),
-    radial-gradient(
-      circle 16px at center,
-      transparent 16px,
-      var(--color-figma-neutral-50) 16px
-    );
-  mask-composite: intersect;
-}
-
-.titleContainer::after {
-  content: '';
-  width: 32px;
-  height: 32px;
-  display: block;
-  position: absolute;
-  background-color: var(--color-figma-neutral-50);
-  right: -32px;
-  top: 0;
-  mask:
-    linear-gradient(
-      to right,
-      var(--color-figma-neutral-50) 50%,
-      transparent 50%
-    ),
-    linear-gradient(to top, transparent 50%, var(--color-figma-neutral-50) 50%),
-    radial-gradient(
-      circle 16px at center,
-      transparent 16px,
-      var(--color-figma-neutral-50) 16px
-    );
-  mask-composite: intersect;
-}
-
-@media (max-width: 1279px) {
-  .titleContainer::before {
-    width: 24px;
-    height: 24px;
-    left: 0;
-    bottom: -24px;
-    mask:
-      linear-gradient(
-        to left,
-        transparent 50%,
-        var(--color-figma-neutral-50) 50%
-      ),
-      linear-gradient(
-        to bottom,
-        var(--color-figma-neutral-50) 50%,
-        transparent 50%
-      ),
-      radial-gradient(
-        circle 12px at center,
-        transparent 12px,
-        var(--color-figma-neutral-50) 12px
-      );
-    mask-composite: intersect;
-  }
-
-  .titleContainer::after {
-    width: 24px;
-    height: 24px;
-    right: -24px;
-    mask:
-      linear-gradient(
-        to right,
-        var(--color-figma-neutral-50) 50%,
-        transparent 50%
-      ),
-      linear-gradient(
-        to top,
-        transparent 50%,
-        var(--color-figma-neutral-50) 50%
-      ),
-      radial-gradient(
-        circle 12px at center,
-        transparent 12px,
-        var(--color-figma-neutral-50) 12px
-      );
-    mask-composite: intersect;
-  }
-}
-
-.descriptionContainer {
-  position: relative;
-}
-
-.descriptionContainer::before {
-  content: '';
-  width: 32px;
-  height: 32px;
-  display: block;
-  position: absolute;
-  background-color: var(--color-figma-neutral-50);
-  left: -32px;
-  bottom: 0;
-  mask:
-    linear-gradient(
-      to left,
-      var(--color-figma-neutral-50) 50%,
-      transparent 50%
-    ),
-    linear-gradient(
-      to bottom,
-      transparent 50%,
-      var(--color-figma-neutral-50) 50%
-    ),
-    radial-gradient(
-      circle 16px at center,
-      transparent 16px,
-      var(--color-figma-neutral-50) 16px
-    );
-  mask-composite: intersect;
-}
-
-.descriptionContainer::after {
-  content: '';
-  width: 32px;
-  height: 32px;
-  display: block;
-  position: absolute;
-  background-color: var(--color-figma-neutral-50);
-  right: 0;
-  top: -32px;
-  mask:
-    linear-gradient(
-      to right,
-      transparent 50%,
-      var(--color-figma-neutral-50) 50%
-    ),
-    linear-gradient(to top, var(--color-figma-neutral-50) 50%, transparent 50%),
-    radial-gradient(
-      circle 16px at center,
-      transparent 16px,
-      var(--color-figma-neutral-50) 16px
-    );
-  mask-composite: intersect;
-}
-
-@media (max-width: 1279px) {
-  .descriptionContainer::before {
-    display: none;
-  }
-
-  .descriptionContainer::after {
-    display: none;
-  }
-}
