+++ src/components/pages/group-tours/config.ts
@@ -1,84 +1,8 @@
-export type TRegion = {
-  id: string
-  name: string
-  countries: TCountry[]
-}
+import type { TSlideContent, TTourDate, TTourData, TTravelerReview } from '@/lib/tours'
+import { REGIONS, SORT_OPTIONS, getCountryCodes, convertCountriesToFilters } from '@/lib/tours'
 
-export type TCountry = {
-  id: string
-  name: string
-  code?: string
-}
-
-export type TSlideContent = {
-  id: number
-  title: string
-  subtitle: string
-  description: string
-}
-
-export type TFilterType = 'country' | 'price' | 'other'
-
-export type TFilter = {
-  id: string
-  label: string
-  type: TFilterType
-}
-
-export type TSelectedFilters = TFilter[]
-
-export type TTourDate = {
-  date: string
-  status: '已成團' | '熱銷中' | '已滿團'
-}
-
-export type TTourData = {
-  id: string
-  title: string
-  subtitle: string
-  description: string
-  price: number
-  tags: string[]
-  dates: TTourDate[]
-  mainImageUrl?: string
-}
-
-// DestinationFilter 搜尋選項
-export const REGIONS: TRegion[] = [
-  {
-    id: 'western-europe',
-    name: '西歐',
-    countries: [
-      { id: 'france', name: '法國', code: 'FR' },
-      { id: 'uk', name: '英國', code: 'GB' },
-      { id: 'germany', name: '德國', code: 'DE' },
-      { id: 'netherlands', name: '荷蘭', code: 'NL' },
-      { id: 'belgium', name: '比利時', code: 'BE' },
-    ],
-  },
-  {
-    id: 'southern-europe',
-    name: '南歐',
-    countries: [
-      { id: 'spain', name: '西班牙', code: 'ES' },
-      { id: 'portugal', name: '葡萄牙', code: 'PT' },
-      { id: 'italy', name: '義大利', code: 'IT' },
-      { id: 'greece', name: '希臘', code: 'GR' },
-    ],
-  },
-  {
-    id: 'northern-europe',
-    name: '北歐',
-    countries: [
-      { id: 'norway', name: '挪威', code: 'NO' },
-      { id: 'sweden', name: '瑞典', code: 'SE' },
-      { id: 'denmark', name: '丹麥', code: 'DK' },
-      { id: 'finland', name: '芬蘭', code: 'FI' },
-    ],
-  },
-]
-
-// Banner 幻燈片內容
+export type { TSelectedFilters, TTourData, TTourDate } from '@/lib/tours'
+export { REGIONS, SORT_OPTIONS, getCountryCodes, convertCountriesToFilters }
 export const SLIDE_CONTENT: TSlideContent[] = [
   {
     id: 1,
@@ -89,59 +13,11 @@ export const SLIDE_CONTENT: TSlideContent[] = [
   },
 ]
 
-export const SORT_OPTIONS = [
-  '價格（低到高）',
-  '價格（高到低）',
-]
-
 export const TOTAL_SLIDES = 6
-
-// 工具函數
-export const getCountryById = (countryId: string): TCountry | null => {
-  for (const region of REGIONS) {
-    const country = region.countries.find((c) => c.id === countryId)
-    if (country) {
-      return country
-    }
-  }
-  return null
-}
-
-export const getCountryCodes = (countryCodes: string[]): string[] => {
-  return countryCodes
-}
-
-export const convertCountriesToFilters = (
-  selectedCountries: string[],
-  regionsData?: Array<{ region: string; countries: Array<{ code: string; nameZh: string }> }>
-): TFilter[] => {
-  if (!regionsData || regionsData.length === 0) {
-    return selectedCountries.map((code) => ({
-      id: code,
-      label: code,
-      type: 'country' as const
-    }))
-  }
-
-  return selectedCountries
-    .map((countryCode) => {
-      for (const region of regionsData) {
-        const country = region.countries.find((c) => c.code === countryCode)
-        if (country) {
-          return {
-            id: countryCode,
-            label: country.nameZh,
-            type: 'country' as const,
-          }
-        }
-      }
-      return null
-    })
-    .filter(Boolean) as TFilter[]
-}
 
 export const convertProductToTourData = (product: {
   id: string
+  isFeatured: boolean
   name: string
   namePrefix?: string
   summary?: string
@@ -150,6 +26,7 @@ export const convertProductToTourData = (product: {
   tags?: string[]
   days?: number
   mainImageUrl?: string
+  feedback?: string | null
   tour?: Array<{
     id: string
     productId: string
@@ -190,15 +67,32 @@ export const convertProductToTourData = (product: {
     return []
   }
 
+  const parseFeedback = (): TTravelerReview | undefined => {
+    if (!product.feedback) return undefined
+
+    try {
+      const feedbackData = JSON.parse(product.feedback)
+      return {
+        author: feedbackData.author || '',
+        avatarUrl: feedbackData.avatarUrl || '',
+      }
+    } catch {
+      return undefined
+    }
+  }
+
   return {
     id: product.id,
+    isFeatured: product.isFeatured,
     title: product.name,
     subtitle: product.namePrefix || '',
     description: product.summary || product.description || '',
     price: product.priceMin,
+    days: product.days,
     tags: product.tags || [],
     dates: convertTourDates(),
     mainImageUrl: product.mainImageUrl || '',
+    travelerReview: parseFeedback(),
   }
 }
