+++ src/lib/tours/constants.ts
@@ -1,0 +1,37 @@
+import type { TRegion } from './types'
+
+export const REGIONS: TRegion[] = [
+  {
+    id: 'western-europe',
+    name: '西歐',
+    countries: [
+      { id: 'france', name: '法國', code: 'FR' },
+      { id: 'uk', name: '英國', code: 'GB' },
+      { id: 'germany', name: '德國', code: 'DE' },
+      { id: 'netherlands', name: '荷蘭', code: 'NL' },
+      { id: 'belgium', name: '比利時', code: 'BE' },
+    ],
+  },
+  {
+    id: 'southern-europe',
+    name: '南歐',
+    countries: [
+      { id: 'spain', name: '西班牙', code: 'ES' },
+      { id: 'portugal', name: '葡萄牙', code: 'PT' },
+      { id: 'italy', name: '義大利', code: 'IT' },
+      { id: 'greece', name: '希臘', code: 'GR' },
+    ],
+  },
+  {
+    id: 'northern-europe',
+    name: '北歐',
+    countries: [
+      { id: 'norway', name: '挪威', code: 'NO' },
+      { id: 'sweden', name: '瑞典', code: 'SE' },
+      { id: 'denmark', name: '丹麥', code: 'DK' },
+      { id: 'finland', name: '芬蘭', code: 'FI' },
+    ],
+  },
+]
+
+export const SORT_OPTIONS = ['價格（低到高）', '價格（高到低）']
